Uploaded all 5 .py files.
These currently-nonfunctional files are all of the program files. They make different types of galaxies, although if you set up a preset to read from you could create a ring galaxy in the Spiral maker and it's very easy to make a lenticular in the elliptical- The only difference is that the lenticular has the same X and Y values and is flat. You can also create a lenticular in the spiral maker, of course.